Update patch set 2
Patch Set 2:

This is a tiny extra bit for the G6 on 17.1.  Without this the device gets a fully random MAC address if misc is wiped.  This extra bit preserves functionality from `hwaddrs` on msm8996-common.  One difference is this sets the second bit to 1 as I believe that is more correct for a randomly generated address.  The other devices will have similar commits.

Figuring you might be interested in this bit too.

Patch-set: 2
CC: Gerrit User 3962 <3962@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Reviewer: Gerrit User 18177 <18177@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>